Move rest of types into feature specific folders
