+++ apps/desktop/src/components/BranchFiles.svelte
@@ -9,7 +9,9 @@
 	import { FileIdSelection } from '$lib/selection/fileIdSelection';
 	import { getContext } from '@gitbutler/shared/context';
 	import type { PatchSeries } from '$lib/branches/branch';
-	import type { DetailedCommit, LocalFile, RemoteFile } from '$lib/vbranches/types';
+	import type { DetailedCommit } from '$lib/commits/commit';
+	import type { RemoteFile } from '$lib/files/file';
+	import type { LocalFile } from '$lib/files/file';
 	import type { Writable } from 'svelte/store';
 
 	interface Props {
