+++ apps/desktop/src/components/BranchFilesHeader.svelte
@@ -3,7 +3,8 @@
 	import { maybeGetContextStore } from '@gitbutler/shared/context';
 	import Badge from '@gitbutler/ui/Badge.svelte';
 	import Checkbox from '@gitbutler/ui/Checkbox.svelte';
-	import type { AnyFile, ConflictEntries } from '$lib/vbranches/types';
+	import type { ConflictEntries } from '$lib/files/conflicts';
+	import type { AnyFile } from '$lib/files/file';
 	import type { Writable } from 'svelte/store';
 
 	interface Props {
