+++ apps/desktop/src/components/CommitContextMenu.svelte
@@ -5,9 +5,9 @@
 	import { BaseBranch } from '$lib/baseBranch/baseBranch';
 	import { BranchStack } from '$lib/branches/branch';
 	import { BranchController } from '$lib/branches/branchController';
+	import { type Commit, type DetailedCommit } from '$lib/commits/commit';
 	import { copyToClipboard } from '$lib/utils/clipboard';
 	import { openExternalUrl } from '$lib/utils/url';
-	import { type Commit, type DetailedCommit } from '$lib/vbranches/types';
 	import { getContext } from '@gitbutler/shared/context';
 
 	interface Props {
