+++ apps/desktop/src/components/CommitDragItem.svelte
@@ -2,8 +2,8 @@
 	import CardOverlay from '$components/CardOverlay.svelte';
 	import Dropzone from '$components/Dropzone.svelte';
 	import { BranchStack } from '$lib/branches/branch';
+	import { Commit, DetailedCommit } from '$lib/commits/commit';
 	import { CommitDragActions, CommitDragActionsFactory } from '$lib/commits/dragActions';
-	import { Commit, DetailedCommit } from '$lib/vbranches/types';
 	import { getContext, maybeGetContextStore } from '@gitbutler/shared/context';
 	import type { Snippet } from 'svelte';
