+++ apps/desktop/src/components/CurrentSeries.svelte
@@ -6,7 +6,7 @@
 	} from '$lib/commits/contexts';
 	import { getColorFromBranchType } from '@gitbutler/ui/utils/getColorFromBranchType';
 	import type { PatchSeries } from '$lib/branches/branch';
-	import type { DetailedCommit } from '$lib/vbranches/types';
+	import type { DetailedCommit } from '$lib/commits/commit';
 	import type { Snippet } from 'svelte';
 
 	interface Props {
