+++ apps/desktop/src/components/EditMode.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import FileContextMenu from '$components/FileContextMenu.svelte';
 	import ScrollableContainer from '$components/ScrollableContainer.svelte';
+	import { Commit } from '$lib/commits/commit';
 	import { CommitService } from '$lib/commits/service';
 	import {
 		conflictEntryHint,
@@ -9,6 +10,7 @@
 		type ConflictEntryPresence,
 		type ConflictState
 	} from '$lib/conflictEntryPresence';
+	import { type RemoteFile } from '$lib/files/file';
 	import { UncommitedFilesWatcher } from '$lib/files/watcher';
 	import { ModeService, type EditModeMetadata } from '$lib/mode/modeService';
 	import { Project } from '$lib/project/project';
@@ -16,7 +18,6 @@
 	import { UserService } from '$lib/user/userService';
 	import { computeFileStatus } from '$lib/utils/fileStatus';
 	import { getEditorUri, openExternalUrl } from '$lib/utils/url';
-	import { Commit, type RemoteFile } from '$lib/vbranches/types';
 	import { getContextStoreBySymbol } from '@gitbutler/shared/context';
 	import { getContext } from '@gitbutler/shared/context';
 	import Badge from '@gitbutler/ui/Badge.svelte';
