+++ apps/desktop/src/components/FileCard.svelte
@@ -5,7 +5,7 @@
 	import { BranchController } from '$lib/branches/branchController';
 	import { ContentSection, HunkSection, parseFileSections } from '$lib/utils/fileSections';
 	import { getContext } from '@gitbutler/shared/context';
-	import type { AnyFile } from '$lib/vbranches/types';
+	import type { AnyFile } from '$lib/files/file';
 
 	interface Props {
 		file: AnyFile;
