+++ apps/desktop/src/components/History.svelte
@@ -5,9 +5,9 @@
 	import ScrollableContainer from '$components/ScrollableContainer.svelte';
 	import SnapshotCard from '$components/SnapshotCard.svelte';
 	import emptyFolderSvg from '$lib/assets/empty-state/empty-folder.svg?raw';
+	import { RemoteFile } from '$lib/files/file';
 	import { HistoryService, createdOnDay } from '$lib/history/history';
 	import { Project } from '$lib/project/project';
-	import { RemoteFile } from '$lib/vbranches/types';
 	import { getContext } from '@gitbutler/shared/context';
 	import Button from '@gitbutler/ui/Button.svelte';
 	import EmptyStatePlaceholder from '@gitbutler/ui/EmptyStatePlaceholder.svelte';
