+++ apps/desktop/src/components/HunkDiff.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import ScrollableContainer from '$components/ScrollableContainer.svelte';
 	import { SelectedOwnership } from '$lib/branches/ownership';
+	import { type Hunk } from '$lib/hunks/hunk';
 	import { create } from '$lib/utils/codeHighlight';
 	import {
 		type ContentSection,
@@ -8,7 +9,6 @@
 		type Line,
 		CountColumnSide
 	} from '$lib/utils/fileSections';
-	import { type Hunk } from '$lib/vbranches/types';
 	import { maybeGetContextStore } from '@gitbutler/shared/context';
 	import Checkbox from '@gitbutler/ui/Checkbox.svelte';
 	import Icon from '@gitbutler/ui/Icon.svelte';
