+++ apps/desktop/src/components/Stack.svelte
@@ -10,6 +10,7 @@
 	import noChangesSvg from '$lib/assets/empty-state/lane-no-changes.svg?raw';
 	import { BranchStack } from '$lib/branches/branch';
 	import { BranchController } from '$lib/branches/branchController';
+	import { DetailedCommit } from '$lib/commits/commit';
 	import { getForgeListingService } from '$lib/forge/interface/forgeListingService';
 	import { Project } from '$lib/project/project';
 	// import { BaseBranch } from '$lib/baseBranch/baseBranch';
@@ -22,7 +23,6 @@
 	// import MergeButton from '$components/MergeButton.svelte';
 	import { intersectionObserver } from '$lib/utils/intersectionObserver';
 	// import * as toasts from '$lib/utils/toasts';
-	import { DetailedCommit } from '$lib/vbranches/types';
 	// import { VirtualBranchService } from '$lib/vbranches/virtualBranch';
 	import { getContext, getContextStore, getContextStoreBySymbol } from '@gitbutler/shared/context';
 	import { persisted } from '@gitbutler/shared/persisted';
