+++ apps/desktop/src/lib/commits/dragActions.ts
@@ -1,7 +1,9 @@
+import { Commit, type DetailedCommit } from './commit';
 import { type BranchStack } from '$lib/branches/branch';
 import { filesToOwnership, filesToSimpleOwnership } from '$lib/branches/ownership';
 import { CommitDropData, FileDropData, HunkDropData } from '$lib/dragging/draggables';
-import { LocalFile, Commit, RemoteFile, type DetailedCommit } from '$lib/vbranches/types';
+import { RemoteFile } from '$lib/files/file';
+import { LocalFile } from '$lib/files/file';
 import type { BranchController } from '$lib/branches/branchController';
 import type { Project } from '$lib/project/project';
