+++ apps/desktop/src/lib/dragging/draggable.ts
@@ -1,6 +1,6 @@
 import { FileDropData, type DropData } from './draggables';
 import { dropzoneRegistry } from './dropzone';
-import { type CommitStatus } from '$lib/vbranches/types';
+import { type CommitStatus } from '$lib/commits/commit';
 import { getFileIcon } from '@gitbutler/ui/file/getFileIcon';
 import { pxToRem } from '@gitbutler/ui/utils/pxToRem';
