+++ apps/desktop/src/lib/files/lock.ts
@@ -1,4 +1,5 @@
-import { HunkLock, type DetailedCommit } from '../vbranches/types';
+import { type DetailedCommit } from '$lib/commits/commit';
+import { HunkLock } from '$lib/hunks/hunk';
 import { uniqeByPropValues, unique } from '$lib/utils/filters';
 
 export function getLockText(hunkLocks: HunkLock | HunkLock[], commits: DetailedCommit[]): string {
