+++ apps/desktop/src/lib/hunks/hunk.ts
@@ -1,0 +1,52 @@
+import { hashCode } from '@gitbutler/ui/utils/string';
+import { Transform, Type } from 'class-transformer';
+import 'reflect-metadata';
+
+export class RemoteHunk {
+	diff!: string;
+	hash?: string;
+	new_start!: number;
+	new_lines!: number;
+	changeType!: ChangeType;
+
+	get id(): string {
+		return hashCode(this.diff);
+	}
+
+	get locked() {
+		return false;
+	}
+}
+export type ChangeType =
+	/// Entry does not exist in old version
+	| 'added'
+	/// Entry is untracked item in workdir
+	| 'untracked'
+	/// Entry does not exist in new version
+	| 'deleted'
+	/// Entry content changed between old and new
+	| 'modified';
+
+export class Hunk {
+	id!: string;
+	diff!: string;
+	@Transform((obj) => {
+		return new Date(obj.value);
+	})
+	modifiedAt!: Date;
+	filePath!: string;
+	hash?: string;
+	locked!: boolean;
+	@Type(() => HunkLock)
+	lockedTo!: HunkLock[];
+	/// Indicates that the hunk depends on multiple branches. In this case the hunk cant be moved or comitted.
+	poisoned!: boolean;
+	changeType!: ChangeType;
+	new_start!: number;
+	new_lines!: number;
+}
+
+export class HunkLock {
+	branchId!: string;
+	commitId!: string;
+}
